Clean up code

+++ proxy/proxy.py
@@ -1,33 +1,22 @@
 #!/usr/bin/env python3
 
 import afl
-afl.init(remote_trace=True)
 import argparse
 import sys
 import os
 import signal
 import socket
-sock = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
-server_address = str(os.getppid())
-mypid = os.getpid()
-try:
-    sock.connect(server_address)
-except FileNotFoundError:
-    print("Instrument won't be sent to AFL")
-    pass
 
 try:
-    from proxy.env import env
+    from env import env
 except:
-    try:
-        from env import env
-    except:
-        env = None
-import frida
+    env = None
 
 
-DESCR = """Frida AFL
+DESCR = """fridaAFL
 """
+
+MAP_SIZE = 64 * 1024
 
 pid = None
 session = None
@@ -66,20 +55,21 @@ def on_message(message, data):
         os._exit(0)
 
 def signal_handler(sig, frame):
-    global args, device, script, session, pid, server_address, mypid
+    global args, device, script, session, pid, server_address
     print('>Catch signal %s, exiting...' % sig)
-    try:
-        os.unlink(server_address)
-    except OSError:
-        if os.path.exists(server_address):
-            raise
+    if sig == 15:
+        try:
+            os.unlink(server_address)
+        except OSError:
+            if os.path.exists(server_address):
+                raise
     if args.s and not args.U:
         print('>Killing', pid)
         os.kill(pid, signal.SIGKILL)
     elif args.s and args.U:
         try:
             device.kill(pid)
-        except frida.ProcessNotFoundError:
+        except:
             print('>Unable to find process :%s' % pid)
             pass
     try:
@@ -87,11 +77,18 @@ def signal_handler(sig, frame):
         session.detach()
     except:
         pass
-    os.kill(mypid, signal.SIGKILL)
+    if sig == 14:
+        try:
+            sock.sendall(b'0'* MAP_SIZE)
+            sock.close()
+        except Exception as e:
+            print("Sock send timeout error: ", e)
+            pass
+    os.kill(os.getpid(), signal.SIGKILL)
 
 
 def main():
-    global args, device, script, session, pid
+    global args, device, script, session, pid, sock, code
     opt = argparse.ArgumentParser(description=DESCR, formatter_class=argparse.RawTextHelpFormatter)
     opt.add_argument('-l', action='store', default='proxy/proxy.js', help='Script filename')
     opt.add_argument('-U', action='store_true', default=False, help='Connect to USB')
@@ -110,6 +107,12 @@ def main():
     with open(args.l) as f:
         code = f.read()
 
+    afl.init(remote_trace=True)
+    fuzz()
+
+def fuzz():
+    global args, device, script, session, pid, sock, code
+    import frida
     try:
         if args.U:
             device = frida.get_usb_device()
@@ -133,6 +136,15 @@ def main():
     script.load()
     signal.signal(signal.SIGINT, signal_handler)
     signal.signal(signal.SIGTERM, signal_handler)
+    signal.signal(signal.SIGALRM, signal_handler)
+
+    sock = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
+    server_address = str(os.getppid())
+    try:
+        sock.connect(server_address)
+    except FileNotFoundError:
+        print("Instrument won't be sent to AFL")
+        pass
 
     try:
         script.exports.fuzzer()
